+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Latest
  - `genie-wish` - removed the wish-action attribute so it no longer has an isolate scope.
  - `ux-lamp` - Added ability to press "tab" on sub context wishes.
  - `ux-lamp` - Altered the template a bit to give more CSS control.
+ - `ux-lamp` - Improved scope.$apply stuff to be safe. Fixed subcontext wish execution.
 
 #0.2.4